vvv.py
from google import genai
from config import GEMINI_API_KEY

client = genai.Client(api_key=GEMINI_API_KEY)

def ask_gemini(prompt: str) -> str:
    response = client.models.generate_content(
        model="models/gemini-flash-lite-latest",
        contents=prompt
    )
    return response.text
